container/src/requirements.txt
docker
paho.mqtt